requirements.txt
Flask==3.0.0
jieba==0.42.1
pypinyin==0.49.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
